remove mutual dependancy between DashManifestModel and DashAdapter by moving getAdaptationForType function in the right class.
