+++ src/streaming/MediaPlayer.js
@@ -221,9 +221,7 @@ function MediaPlayer() {
 
         adapter = DashAdapter(context).getInstance();
         dashManifestModel = DashManifestModel(context).getInstance({
-            mediaController: mediaController,
             timelineConverter: timelineConverter,
-            adapter: adapter,
             errHandler: errHandler
         });
         manifestModel = ManifestModel(context).getInstance();
@@ -239,6 +237,7 @@ function MediaPlayer() {
         });
 
         adapter.setConfig({
+            mediaController: mediaController,
             dashManifestModel: dashManifestModel
         });
 
@@ -2750,7 +2749,6 @@ function MediaPlayer() {
             mediaPlayerModel: mediaPlayerModel,
             metricsModel: metricsModel,
             dashMetrics: dashMetrics,
-            dashManifestModel: dashManifestModel,
             manifestModel: manifestModel,
             videoModel: videoModel,
             adapter: adapter
