+++ src/streaming/controllers/AbrController.js
@@ -76,7 +76,6 @@ function AbrController() {
         elementWidth,
         elementHeight,
         manifestModel,
-        dashManifestModel,
         adapter,
         videoModel,
         mediaPlayerModel,
@@ -185,9 +184,6 @@ function AbrController() {
         if (config.dashMetrics) {
             dashMetrics = config.dashMetrics;
         }
-        if (config.dashManifestModel) {
-            dashManifestModel = config.dashManifestModel;
-        }
         if (config.adapter) {
             adapter = config.adapter;
         }
@@ -265,7 +261,7 @@ function AbrController() {
         if (!bitrateDict.hasOwnProperty(type)) {
             if (ratioDict.hasOwnProperty(type)) {
                 const manifest = manifestModel.getValue();
-                const representation = dashManifestModel.getAdaptationForType(manifest, 0, type).Representation;
+                const representation = adapter.getAdaptationForType(manifest, 0, type).Representation;
 
                 if (Array.isArray(representation)) {
                     const repIdx = Math.max(Math.round(representation.length * ratioDict[type]) - 1, 0);
@@ -689,7 +685,7 @@ function AbrController() {
         }
 
         const manifest = manifestModel.getValue();
-        const representation = dashManifestModel.getAdaptationForType(manifest, 0, type).Representation;
+        const representation = adapter.getAdaptationForType(manifest, 0, type).Representation;
         let newIdx = idx;
 
         if (elementWidth > 0 && elementHeight > 0) {
